+++ src/Config/ConfigFile/Reader/ConfigFileReader.php
@@ -22,7 +22,7 @@ use Puli\RepositoryManager\InvalidConfigException;
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-interface ConfigFileReaderInterface
+interface ConfigFileReader
 {
     /**
      * Reads a configuration file.