+++ src/Package/PackageFile/RootPackageFile.php
@@ -197,7 +197,7 @@ class RootPackageFile extends PackageFile
      * Sets the plugin classes.
      *
      * The plugin classes must be fully-qualified class names that implement
-     * {@link \Puli\RepositoryManager\Plugin\PluginInterface}. If a class is not
+     * {@link \Puli\RepositoryManager\Plugin\ManagerPlugin}. If a class is not
      * found or does not implement that interface, an exception is thrown.
      *
      * The plugin classes must not have required parameters in their constructor
@@ -208,9 +208,8 @@ class RootPackageFile extends PackageFile
      *
      * @param string[] $pluginClasses The fully qualified plugin class names.
      *
-     * @throws InvalidConfigException If a class is not found, is not a class,
-     *                                does not implement
-     *                                {@link \Puli\RepositoryManager\Plugin\PluginInterface}
+     * @throws InvalidConfigException If the class is not found, is not a class,
+     *                                does not implement {@link ManagerPlugin}
      *                                or has required constructor parameters.
      */
     public function setPluginClasses(array $pluginClasses)
@@ -226,8 +225,8 @@ class RootPackageFile extends PackageFile
      * Adds a plugin class.
      *
      * The plugin class must be a fully-qualified class name that implements
-     * {@link \Puli\RepositoryManager\Plugin\PluginInterface}. If the class is not
-     * found or does not implement that interface, an exception is thrown.
+     * {@link ManagerPlugin}. If the class is not found or does not implement
+     * that interface, an exception is thrown.
      *
      * The plugin class must not have required parameters in its constructor
      * so that it can be successfully instantiate. If the constructor has
@@ -238,8 +237,7 @@ class RootPackageFile extends PackageFile
      * @param string $pluginClass The fully qualified plugin class name.
      *
      * @throws InvalidConfigException If the class is not found, is not a class,
-     *                                does not implement
-     *                                {@link \Puli\RepositoryManager\Plugin\PluginInterface}
+     *                                does not implement {@link ManagerPlugin}
      *                                or has required constructor parameters.
      */
     public function addPluginClass($pluginClass)
@@ -267,9 +265,9 @@ class RootPackageFile extends PackageFile
             ));
         }
 
-        if (!$reflClass->implementsInterface('\Puli\RepositoryManager\Plugin\PluginInterface')) {
+        if (!$reflClass->implementsInterface('\Puli\RepositoryManager\Plugin\ManagerPlugin')) {
             throw new InvalidConfigException(sprintf(
-                'The plugin class %s must implement \Puli\RepositoryManager\Plugin\PluginInterface.',
+                'The plugin class %s must implement ManagerPlugin.',
                 $pluginClass
             ));
         }