+++ src/Package/PackageFile/Writer/PackageJsonWriter.php
@@ -28,7 +28,7 @@ use Webmozart\PathUtil\Path;
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class PackageJsonWriter implements PackageFileWriterInterface
+class PackageJsonWriter implements PackageFileWriter
 {
     /**
      * The default order of the keys in the written package file.