+++ tests/Package/PackageFile/Fixtures/TestPluginInterface.php
@@ -11,12 +11,12 @@
 
 namespace Puli\RepositoryManager\Tests\Package\PackageFile\Fixtures;
 
-use Puli\RepositoryManager\Plugin\PluginInterface;
+use Puli\RepositoryManager\Plugin\ManagerPlugin;
 
 /**
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-interface TestPluginInterface extends PluginInterface
+interface TestPluginInterface extends ManagerPlugin
 {
 }