+++ tests/Package/PackageFile/Fixtures/TestPluginWithoutNoArgConstructor.php
@@ -12,13 +12,13 @@
 namespace Puli\RepositoryManager\Tests\Package\PackageFile\Fixtures;
 
 use Puli\RepositoryManager\Environment\ProjectEnvironment;
-use Puli\RepositoryManager\Plugin\PluginInterface;
+use Puli\RepositoryManager\Plugin\ManagerPlugin;
 
 /**
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class TestPluginWithoutNoArgConstructor implements PluginInterface
+class TestPluginWithoutNoArgConstructor implements ManagerPlugin
 {
     public function __construct($arg)
     {